auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
python-jose[cryptography]
pydantic